non darcy flow
tons of output of an old air spring model with darcy flow.

Abedi says darcy flow is not accurate to model bag motion, and that it
serves as an overestimate. There is now implemented non darcy flow
through the forchheimer model (linear + quadratic). This should more
accurately model flow, and it does. Next step is to find realistic
values for A and B in the forchheimer model.